+++ data/hemisphere.js
@@ -1,0 +1,38 @@
+//used for summer/winter things
+module.exports = {
+  south: {
+    'America/Argentina': true,
+    'America/Lima': true,
+    'America/La_Paz': true,
+    'America/Manaus': true,
+    'America/Asuncion': true,
+    'America/Buenos_Aires': true,
+    'America/Montevideo': true,
+    'America/Cordoba': true,
+    'America/Stanley': true,
+    'Brazil/Acre': true,
+
+    'Africa/Luanda': true,
+    'Africa/Windhoek': true,
+    'Africa/Maseru': true,
+    'Africa/Gaborone': true,
+    'Africa/Maputo': true,
+    'Africa/Mbabane': true,
+    'Africa/Lusaka': true,
+    'Africa/Kinshasa': true,
+    'Africa/Johannesburg': true,
+    'Africa/Lubumbashi': true,
+
+    'Asia/Dili': true,
+    'Asia/Makassar': true,
+    'Asia/Jakarta': true,
+    'Asia/Singapore': true,
+    'Asia/Kuala_Lumpur': true,
+    'Chile/EasterIsland': true,
+    'Pacific/Apia': true,
+    'Pacific/Chatham': true,
+    'Pacific/Easter': true,
+
+    'Indian/Reunion': true
+  }
+};